+++ CHANGELOG.md
@@ -17,6 +17,14 @@ y este proyecto utiliza [SemVer](https://semver.org/lang/es/).
 
 ---
 
+## [1.12.1] - 2025-10-05
+
+### Added
+- Added youtube api usage metrics per channel
+
+### Fixed
+- Fixed cron not using batch fetch
+
 ## [1.12.0] - 2025-10-04
 
 ### Changed